Fixed a miscalculated beam_frqs array
We double-counted 151 MHz and neglected to include 250 MHz
